requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-login==0.6.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
python-dotenv==1.0.0
joblib==1.3.2
gunicorn==21.2.0
mysqlclient==2.2.0
werkzeug==2.3.7
